Adds video converter and downloader
Adds a new and independent tool to help users converting .mp4 files to .ogg.
This could be used both in local or remote mode. If local is selected, only the
conversion will be done. If remote mode is selected, this same tool will make
the download of video and image files by the user using the given credentials.
Also the compose file is modified to accept to new arguments from the env file,
that will be used on the database dockerfile to do the download of the sql file
in the same way of the previous tool, using the user credentials.
The readme is also updated to include a explanation on how to install and use
the tools.