Fixes backend startup error due to missing configuration property
The backend was updated since piba-docker was last reviewed to support
storing the files originally uploaded by users to a directory, and
requires such a directory to be explicitly configured.

For the purposes of piba-docker, let's set such property to the special
value "DISABLED", which disables this new feature, unnecessary for
reproducibility, that requires its directory to be created in the
media volume.