Prevents backend container from starting before database container
Currently, the backend container starts after the database container,
but there is no guarantee that the database container finishes database
initialization before the backend container attempts to connect to its
database.

To fix that race condition, define a health check for the database
container that checks whether the database daemon is accepting
connections, and instruct the backend container to wait until it passes.