Fixes backend failing to launch under some vague environment conditions
All files under /opt/jboss/wildfly should belong to the jboss user and
group, but when deploying the backend on a different development
environment this wasn't the case, which caused errors when the
application container tried to manipulate log files.

Fix that problem by recursively setting ownership for every file related
to the application container, which is a more heavy-handed approach, but
should work fine on such cases.